refactor: dependencies, lint, 화면 텍스트 수정, OnboardingLayout designSystem으로 편입
